Cache specific types for loading
This makes loading from the cache resilient to data struture changes, as
these cache types do not implement the deserialization error skipping
logic that the main types do. This means if we change the Music or
Library structs to add/remove fields, loading from cache will properly
fail resulting in the application fetching from Jellyfin again which is
what we want, replacing the cache with the properly typed new structure.